intial move around
